+++ src/components/@common/Buttons/Button/Button.stories.tsx
@@ -1,6 +1,6 @@
 import {ComponentMeta, ComponentStory} from '@storybook/react';
 
-import Button from 'components/@common/Button/Button';
+import Button from 'components/@common/Buttons/Button';
 
 const meta = {
   title: 'components/Button',
