+++ src/components/@common/Buttons/SocialLoginButton/SocialLoginButton.stories.tsx
@@ -1,6 +1,6 @@
 import {ComponentMeta, ComponentStory} from '@storybook/react';
 
-import SocialLoginButton from 'components/@common/Button/SocialLoginButton';
+import SocialLoginButton from 'components/@common/Buttons/SocialLoginButton';
 
 const meta = {
   title: 'components/Button',
